Move all to use new useSearch
