+++ src/shared/hooks/useSearch.ts
@@ -7,11 +7,14 @@
  * returned data from whatever fetched data you fetch.
  */
 import { useState, useCallback, useEffect, useMemo } from "react";
+import { DocumentNode } from "graphql";
+import { QueryHookOptions, useQuery } from "@apollo/react-hooks";
 
 import { useSearchParams } from "src/shared/hooks/useSearchParams";
 import { RESULTS_PER_PAGE } from "src/shared/constants/search";
-import { analytics } from "src/shared/utils/analytics";
 import { IGenericCardItem } from "src/shared/constants/card";
+
+import { analytics } from "src/shared/utils/analytics";
 
 /*******************************************************************
  *                             **Types**                           *
@@ -22,7 +25,7 @@ export enum SearchState {
   LOADING,
   ERROR,
   RESULTS,
-  RESULTS_LOADING, // results exist, but is searching
+  RESULTS_LOADING, // prior results, but is currently loading
   NO_RESULTS,
   NO_MORE_RESULTS,
 }
@@ -30,120 +33,42 @@ export enum SearchState {
 /*******************************************************************
  *                              **Hook**                           *
  *******************************************************************/
-export const useSearch = () => {
-  const { searchQuery, searchType, setSearchQuery } = useSearchParams();
-
-  /**
-   * Every time the search query is updated, update the search results.
-   * Assuming that updating the search query is debounced or throttled
-   * so it doesn't get updated too ofte, this API call is not made excessively.
-   */
-  const [page, setPage] = useState(1);
-  const [isNewSearch, setIsNewSearch] = useState(false); // whether a search is completely new or just another page of the current search
-  const [isEndOfResults, setIsEndOfResults] = useState(false); // whether the end of
-  const isInitialSearch = useMemo(() => searchQuery === undefined, [
-    searchQuery,
-  ]); // whether a user has not yet searched for the first time
-  const [isDataLoaded, setIsDataLoaded] = useState(false); // whether data is loaded and ready
-
-  const onNewSearch = useCallback(
-    (newVal?: string) => {
-      if (newVal !== undefined && newVal !== searchQuery) {
-        analytics.event({
-          category: "Search",
-          action: "Started new search",
-          label: newVal,
-        });
-
-        setPage(1); // reset pagination
-        setIsNewSearch(true);
-
-        // prep for search
-        setIsDataLoaded(false);
-        // setIsInitialSearch(false);
-
-        // perform the new search
-        setSearchQuery(newVal);
-      }
-    },
-    [searchQuery, setSearchQuery]
-  );
-
-  const onNextBatchSearch = useCallback(() => {
-    analytics.event({
-      category: "Search",
-      action: "Loaded more search results",
-    });
-
-    // increment pagination
-    setPage(prevPage => prevPage + 1);
-    setIsNewSearch(false);
-
-    // prep for search
-    setIsDataLoaded(false);
-    // setIsInitialSearch(false);
-  }, []);
-
+export const useSearch = <TData>(
+  query: DocumentNode,
+  options: QueryHookOptions<TData>,
+  transformData: (data?: TData) => IGenericCardItem[]
+) => {
+  const { searchQuery, setSearchQuery } = useSearchParams();
   const [searchState, setSearchState] = useState(SearchState.INITIAL);
 
-  return {
-    // filters
-    searchQuery,
-    searchType,
+  const [page, setPage] = useState(1); // most recent page fetched for query
+  const [isNewSearch, setIsNewSearch] = useState(false); // whether a search is completely new or just another page of the current search
+  const [isEndOfResults, setIsEndOfResults] = useState(false); // whether there are more results or not
+  const [isDataLoaded, setIsDataLoaded] = useState(false); // whether data is loaded and ready
 
-    // search info
-    page,
-    searchState,
-
-    // callbacks
-    onNewSearch,
-    onNextBatchSearch,
-
-    // part of searchconfig
-    isNewSearch,
-    isEndOfResults,
-    isInitialSearch,
-    isDataLoaded,
-
-    setIsEndOfResults,
-    setIsDataLoaded,
-    setSearchState,
-  };
-};
-
-interface UseSearchResultsConfig<T> {
-  data?: T;
-  error: boolean;
-  loading: boolean;
-  isEndOfResults: boolean;
-  isNewSearch: boolean;
-  isInitialSearch: boolean;
-  isDataLoaded: boolean;
-  setIsDataLoaded: React.Dispatch<React.SetStateAction<boolean>>;
-  setIsEndOfResults: React.Dispatch<React.SetStateAction<boolean>>;
-  setSearchState: React.Dispatch<React.SetStateAction<SearchState>>;
-}
-
-export const useSearchAfter = <T>(
-  {
-    data,
-    error,
-    loading,
-
-    isNewSearch,
-    isEndOfResults,
-    isInitialSearch,
-    isDataLoaded,
-
-    setIsDataLoaded,
-    setIsEndOfResults,
-    setSearchState,
-  }: UseSearchResultsConfig<T>,
-  transformData: (data?: T) => IGenericCardItem[]
-) => {
   /**
-   * After new data is fetched, build the list of new search results.
-   * Update other invariants accordingly.
+   * *Perform the search* and grab loading/error state and the
+   * data resulting from the search.
+   */
+  const shouldSkipSearch = useMemo(
+    // do not make an API call if search query is empty (on initial load)
+    () => searchState === SearchState.RESULTS || options.skip,
+    [options.skip, searchState]
+  );
+  const { loading, error, data } = useQuery<TData>(query, {
+    ...options,
+    variables: {
+      query: searchQuery || "",
+      offset: (page - 1) * RESULTS_PER_PAGE,
+      limit: RESULTS_PER_PAGE,
+      ...(options.variables || {}),
+    },
+    skip: shouldSkipSearch,
+  });
+
+  /**
+   * After new data is fetched, *build list of new results*.
+   * Update other info accordingly.
    */
   const [searchResults, setSearchResults] = useState<IGenericCardItem[]>([]);
   useEffect(() => {
@@ -160,11 +85,8 @@ export const useSearchAfter = <T>(
         }
       }
 
-      /**
-       * Check whether or not there are more results to be fetched.
-       * If not, indicate that so we can display it to the user and
-       * prevent further API calls with the same query.
-       */
+      // check whether or not there are more results to be fetched, so we can display
+      // no more results as well as prevent further API calls with the same query.
       if (newResults.length < RESULTS_PER_PAGE) {
         setIsEndOfResults(true);
       } else {
@@ -183,13 +105,16 @@ export const useSearchAfter = <T>(
     transformData,
   ]);
 
+  /**
+   * *Track the state of searching*.
+   */
   useEffect(() => {
     let newState = SearchState.INITIAL;
 
     if (error) newState = SearchState.ERROR;
     else if (searchResults.length === 0 && loading)
       newState = SearchState.LOADING;
-    else if (searchResults.length === 0 && !isInitialSearch)
+    else if (searchResults.length === 0 && searchQuery !== undefined)
       newState = SearchState.NO_RESULTS;
     else if (isEndOfResults) newState = SearchState.NO_MORE_RESULTS;
     else if (searchResults.length > 0 && (!isDataLoaded || loading))
@@ -201,11 +126,58 @@ export const useSearchAfter = <T>(
     error,
     isDataLoaded,
     isEndOfResults,
-    isInitialSearch,
     loading,
+    searchQuery,
     searchResults.length,
     setSearchState,
   ]);
 
-  return searchResults;
+  /**
+   * Create *callbacks for triggering search*.
+   */
+  const triggerSearchNew = useCallback(
+    (newVal?: string) => {
+      if (newVal !== undefined && newVal !== searchQuery) {
+        analytics.event({
+          category: "Search",
+          action: "Started new search",
+          label: newVal,
+        });
+
+        setPage(1); // reset pagination
+        setIsNewSearch(true);
+
+        // indicate that a search has started
+        setIsDataLoaded(false);
+
+        // perform the new search
+        setSearchQuery(newVal);
+      }
+    },
+    [searchQuery, setSearchQuery]
+  );
+
+  const triggerSearchNextBatch = useCallback(() => {
+    analytics.event({
+      category: "Search",
+      action: "Loaded more search results",
+    });
+
+    // increment page since we fetched a page
+    setPage(prevPage => prevPage + 1);
+    setIsNewSearch(false);
+
+    // indicate that a search has started
+    setIsDataLoaded(false);
+  }, []);
+
+  return {
+    // search info
+    searchState,
+    searchResults,
+
+    // callbacks
+    triggerSearchNew,
+    triggerSearchNextBatch,
+  };
 };
